requirements.txt
python-telegram-bot==20.7
PyYAML>=6.0
requests>=2.31.0 